requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
Pillow==10.4.0
geopy==2.4.1
openai==1.107.0
spotipy==2.25.1
APScheduler==3.11.1
pytz==2025.2
uwuify==1.3.1
HomeAssistant-API